+++ PoliNetworkBot_CSharp/Code/Utils/Groups.cs
@@ -17,7 +17,7 @@ namespace PoliNetworkBot_CSharp.Code.Utils
             return SqLite.ExecuteSelect(q1);
         }
 
-        internal static async Task<SuccessWithException> CheckIfAdminAsync(int userId, string username, long chatId,
+        internal static async Task<SuccessWithException> CheckIfAdminAsync(long userId, string username, long chatId,
             TelegramBotAbstract telegramBotAbstract)
         {
             if (GlobalVariables.Creators.Contains(username))
